bugfix; add sort() to indexer find() method; add _lt_ method to spangroup and span; move indexer to its own file
